Ensure that all dependencies are defined
Tried setting up a clean gemset to validate if all dependencies needed
by monk-datamapper-skeleton are listed in dependencies.  Added missing
dependencies in the process.  Also replaced 'data_mapper' metagem
requirement with the actual gems it refers to, as we could not define
data_mapper in dependencies directly.

Unfortunately, we could not specify do_postgres as a dependencies as
I think this requires native compilation which only the gem command can
do.  Thus we leave do_postgres out of the dependencies.

Signed-off-by: Ealden Esto E. Escanan <ealden@teamcodeflux.com>